+++ Nagisa/XamlResources/CustomConverters.xaml.h
@@ -70,24 +70,6 @@ namespace Nagisa
 
 	[Windows::UI::Xaml::Data::Bindable]
 	[Windows::Foundation::Metadata::WebHostHidden]
-	public ref class StorageFileToFileNameConverter sealed : IValueConverter
-	{
-	public:
-		// Inherited via IValueConverter
-		virtual Object^ Convert(
-			Object^ value,
-			TypeName targetType,
-			Object^ parameter,
-			String^ language);
-		virtual Object^ ConvertBack(
-			Object^ value,
-			TypeName targetType,
-			Object ^parameter,
-			String^ language);
-	};
-
-	[Windows::UI::Xaml::Data::Bindable]
-	[Windows::Foundation::Metadata::WebHostHidden]
 	public ref class TaskListEmptyToVisibilityConverter sealed
 		: IValueConverter
 	{